Feat, specified mounting, mounted, isMounted methods. mounting method is a special method called if needed before render
