+++ src/types/component.ts
@@ -2,4 +2,7 @@ import { Children, VNode } from "./vnode";
 
 export default interface Component {
   render(props: VNode["props"]) : VNode
+  mounting() : void
+  isMounted() : boolean
+  mount() : void
 }
